basic meta setup
