+++ guide/app/routes/_guide.index.tsx
@@ -1,37 +1,35 @@
-import { type LoaderArgs, json } from '@remix-run/cloudflare';
+import {
+	type LoaderArgs,
+	type HeadersFunction,
+	json,
+} from '@remix-run/cloudflare';
 import { useLoaderData } from '@remix-run/react';
 import { getBranch } from '~/context';
 import { parse } from '~/markdoc.server';
 import { getGitHubReadme } from '~/octokit';
 import { Markdown, Sandbox } from '~/components';
+import { getIntroduction } from '~/format';
 
-function getIntroduction(content: string) {
-	const lines = [];
-
-	for (const line of content.split('\n')) {
-		if (line.startsWith('## ')) {
-			break;
-		}
-
-		if (line.startsWith('# ')) {
-			lines.push('# Conform');
-		} else {
-			lines.push(line);
-		}
-	}
-
-	return lines.join('\n');
-}
+export const headers: HeadersFunction = ({ loaderHeaders }) => {
+	return loaderHeaders;
+};
 
 export async function loader({ context }: LoaderArgs) {
 	const branch = getBranch(context);
 	const readme = await getGitHubReadme(branch);
 	const introduction = getIntroduction(atob(readme.content));
 
-	return json({
-		src: `edmundhung/conform/tree/${branch}/examples/basics`,
-		content: parse(introduction),
-	});
+	return json(
+		{
+			src: `edmundhung/conform/tree/${branch}/examples/basics`,
+			content: parse(introduction),
+		},
+		{
+			headers: {
+				'Cache-Control': 'public, max-age=300',
+			},
+		},
+	);
 }
 
 export default function Index() {
