+++ guide/public/_routes.json
@@ -1,5 +1,5 @@
 {
 	"version": 1,
 	"include": ["/*"],
-	"exclude": ["/build/*"]
+	"exclude": ["/build/*", "/fonts/*"]
 }
